Add files via upload
This is also an example of an MLP in which we use the MNIST dataset to train our model and calculate accuracy. However, this is not the best model; we can still improve accuracy by increasing the number of epochs. It's important to note that simply increasing the number of epochs does not always yield better results, as I will demonstrate in the next file